+++ spec/requests/flows/authorization_code_spec.rb
@@ -526,5 +526,37 @@ describe "Authorization Code Flow" do
       should_have_json "error", "invalid_grant"
       should_have_json "error_description", translated_error_message(:invalid_grant)
     end
+
+    context 'with a token_creation_wrapper' do
+      it 'wraps token creation' do
+        # Creating the wrapper here so we have access to wrapper_count
+        wrapper_count = 0
+        wrapper = ->(&block) do
+          wrapper_count += 1
+          block.call(repeat_find: false)
+        end
+        config_is_set(:token_creation_wrapper, wrapper)
+
+        authorization_code = Doorkeeper::AccessGrant.first.token
+        post token_endpoint_url(code: authorization_code, client: @client)
+
+        should_have_json "access_token", Doorkeeper::AccessToken.first.token
+        should_have_json "token_type", "Bearer"
+        should_have_json_within "expires_in", Doorkeeper::AccessToken.first.expires_in, 1
+        expect(wrapper_count).to eq 1
+
+        authorization_code_exists application: @client,
+          resource_owner_id: resource_owner.id,
+          resource_owner_type: resource_owner.class.name
+
+        authorization_code = Doorkeeper::AccessGrant.last.token
+        post token_endpoint_url(code: authorization_code, client: @client)
+
+        should_have_json "access_token", Doorkeeper::AccessToken.last.token
+        should_have_json "token_type", "Bearer"
+        should_have_json_within "expires_in", Doorkeeper::AccessToken.last.expires_in, 1
+        expect(wrapper_count).to eq 2
+      end
+    end
   end
 end